Extract kit_scan and kitmanager tests from unit test collection
