+++ test/unit_tests/extension.test.ts
@@ -11,323 +11,16 @@ import * as vscode from 'vscode';
 chai.use(chaiAsPromised);
 
 import {expect} from 'chai';
-import * as sinon from 'sinon';
-
-import * as json5 from 'json5';
-import * as ajv from 'ajv';
-
 import * as api from '../../src/api';
 import * as compdb from '../../src/compdb';
 import {CMakeCache} from '../../src/cache';
 import * as diags from '../../src/diagnostics';
-import * as kit from '../../src/kit';
-import {fs} from '../../src/pr';
 import {OutputConsumer} from '../../src/proc';
-import * as state from '../../src/state';
 import * as util from '../../src/util';
 
 
 const here = __dirname;
-function getTestRootFilePath(filename: string): string { return path.normalize(path.join(here, '../../..', 'test', filename)); }
-
-
 function getTestResourceFilePath(filename: string): string { return path.normalize(path.join(here, '../../../test/unit_tests', filename)); }
-
-function getResourcePath(filename: string): string { return path.normalize(path.join(here, '../../..', filename)); }
-
-function getPathWithoutCompilers() {
-  if (process.arch == "win32") {
-    return "C:\\TMP"
-  } else {
-    return "/tmp"
-  }
-}
-
-suite('Kits test', async () => {
-  const fakebin = getTestRootFilePath('fakebin');
-  test('Detect system kits never throws',
-       async () => {
-         // Don't care about the result, just check that we don't throw during the test
-         await expect(kit.scanForKits()).to.eventually.not.be.rejected;
-       })
-      // Compiler detection can run a little slow
-      .timeout(12000);
-
-  test('Detect a GCC compiler file', async () => {
-    const compiler = path.join(fakebin, 'gcc-42.1');
-    const compkit = await kit.kitIfCompiler(compiler);
-    expect(compkit).to.not.be.null;
-    expect(compkit!.compilers).has.property('C').equal(compiler);
-    expect(compkit!.compilers).to.not.have.property('CXX');
-    expect(compkit!.name).to.eq('GCC 42.1');
-  });
-
-  test('Detect a Clang compiler file', async () => {
-    const compiler = path.join(fakebin, 'clang-0.25');
-    const compkit = await kit.kitIfCompiler(compiler);
-    expect(compkit).to.not.be.null;
-    expect(compkit!.compilers).has.property('C').eq(compiler);
-    expect(compkit!.compilers).to.not.have.property('CXX');
-    expect(compkit!.name).to.eq('Clang 0.25');
-  });
-
-  test('Detect non-compiler program', async () => {
-    const program = path.join(fakebin, 'gcc-666');
-    const nil = await kit.kitIfCompiler(program);
-    expect(nil).to.be.null;
-  });
-
-  test('Detect non existing program', async () => {
-    const program = path.join(fakebin, 'unknown');
-    const nil = await kit.kitIfCompiler(program);
-    expect(nil).to.be.null;
-  });
-
-  test('Scan non exisiting dir for kits', async() => {
-    const kits = await kit.scanDirForCompilerKits('');
-    expect(kits.length).to.eq(0);
-  });
-
-  suite('Scan directory', async () => {
-    let path_with_compilername = '';
-    setup(async () => { path_with_compilername = path.join(fakebin, "gcc-4.3.2"); });
-    teardown(async () => {
-      if (await fs.exists(path_with_compilername)) {
-        await fs.rmdir(path_with_compilername)
-      }
-    });
-    test('Scan folder with compiler name', async () => {
-      fs.mkdir(path_with_compilername)
-      // Scan the directory with fake compilers in it
-      const kits = await kit.scanDirForCompilerKits(fakebin);
-      expect(kits.length).to.eq(2);
-    });
-
-    test('Scan file with compiler name', async () => {
-      await fs.writeFile(path_with_compilername, '')
-      // Scan the directory with fake compilers in it
-      const kits = await kit.scanDirForCompilerKits(fakebin);
-      expect(kits.length).to.eq(2);
-    });
-  });
-
-  suite('Rescan kits', async () => {
-    let km: kit.KitManager;
-    let path_rescan_kit = getTestResourceFilePath('rescan_kit.json');
-    let sandbox: sinon.SinonSandbox;
-    let path_backup: string|undefined;
-    setup(async () => {
-      sandbox = sinon.sandbox.create();
-      let stateMock = sandbox.createStubInstance(state.StateManager);
-      sandbox.stub(stateMock, 'activeKitName').get(function() { return null; }).set(function() {});
-      km = new kit.KitManager(stateMock, path_rescan_kit);
-
-      // Mock showInformationMessage to suppress needed user choice
-      sandbox.stub(vscode.window, "showInformationMessage")
-          .callsFake(function() { return {title : "No", isCloseAffordance : true, doOpen : false}; });
-
-      path_backup = process.env.PATH;
-    });
-    teardown(async () => {
-      sandbox.restore();
-      if (await fs.exists(path_rescan_kit)) {
-        await fs.rmdir(path_rescan_kit)
-      }
-      process.env.PATH = path_backup;
-    });
-
-    async function readValidKitFile(file_path: string): Promise<any[]> {
-      const rawKitsFromFile = (await fs.readFile(file_path, 'utf8'));
-      expect(rawKitsFromFile.length).to.be.not.eq(0);
-
-      let kitFile = json5.parse(rawKitsFromFile);
-
-      const schema = json5.parse(await fs.readFile(getResourcePath('schemas/kits-schema.json'), 'utf8'));
-      const validator = new ajv({allErrors : true, format : 'full'}).compile(schema);
-      expect(validator(kitFile)).to.be.true;
-
-      return kitFile;
-    }
-
-    test('init kit file creation no compilers in path', async () => {
-      process.env['PATH'] = getPathWithoutCompilers();
-
-      await km.initialize();
-
-      const newKitFileExists = await fs.exists(path_rescan_kit);
-      expect(newKitFileExists).to.be.true;
-    }).timeout(10000);
-
-    test('check valid kit file for test system compilers', async () => {
-      await km.initialize();
-
-      await readValidKitFile(path_rescan_kit);
-    }).timeout(30000);
-
-    test('check empty kit file no compilers in path', async () => {
-      process.env['PATH'] = getPathWithoutCompilers();
-
-      await km.initialize();
-
-      let kitFile = await readValidKitFile(path_rescan_kit);
-      let nonVSKits = kitFile.filter((item) => {return item.visualStudio == null});
-      expect(nonVSKits.length).to.be.eq(0);
-    }).timeout(10000);
-
-    // Fails because PATH is tried to split but a empty path is not splitable
-    test.skip('check empty kit file', async () => {
-      process.env['PATH'] = '';
-
-      await km.initialize();
-
-      const newKitFileExists = await fs.exists(path_rescan_kit);
-      expect(newKitFileExists).to.be.true;
-    });
-
-    test('check fake compilers in kit file', async () => {
-      process.env['PATH'] = getTestRootFilePath("fakebin");
-
-      await km.initialize();
-
-      let kitFile = await readValidKitFile(path_rescan_kit);
-      let nonVSKits = kitFile.filter((item) => {return item.visualStudio == null});
-      expect(nonVSKits.length).to.be.eq(2);
-    }).timeout(10000);
-
-    test('check check combination of scan and old kits', async () => {
-      process.env['PATH'] = getTestRootFilePath("fakebin");
-      await fs
-          .copyFile(getTestResourceFilePath('test_kit.json'), path_rescan_kit)
-
-              await km.initialize();
-      await km.rescanForKits()
-
-      let names = km.kits.map((item) => {return item.name});
-
-      expect(names).to.contains("CompilerKit 1");
-      expect(names).to.contains("CompilerKit 2");
-      expect(names).to.contains("CompilerKit 3 with PreferedGenerator");
-      expect(names).to.contains("ToolchainKit 1");
-      expect(names).to.contains("VSCode Kit 1");
-      expect(names).to.contains("VSCode Kit 2");
-      expect(names).to.contains("Clang 0.25");
-      expect(names).to.contains("GCC 42.1");
-    }).timeout(10000);
-  });
-
-  suite('GUI test', async () => {
-    let km: kit.KitManager;
-    let gui_sandbox: sinon.SinonSandbox;
-    setup(async () => {
-      gui_sandbox = sinon.sandbox.create();
-      let stateMock = gui_sandbox.createStubInstance(state.StateManager);
-      sinon.stub(stateMock, 'activeKitName').get(function() { return null; }).set(function() {});
-      const kit_file = getTestResourceFilePath('test_kit.json');
-      km = new kit.KitManager(stateMock, kit_file);
-    });
-    teardown(async () => { gui_sandbox.restore(); });
-
-    test('KitManager tests opening of kit file', async () => {
-      let text: vscode.TextDocument|undefined;
-      gui_sandbox.stub(vscode.window, "showTextDocument").callsFake(function(textDoc) {
-        text = textDoc;
-        return {document : text};
-      });
-      await km.initialize();
-
-      const editor = await km.openKitsEditor();
-
-      expect(text).to.be.not.undefined;
-      if (text != undefined) {
-        const rawKitsFromFile = (await fs.readFile(getTestResourceFilePath('test_kit.json'), 'utf8'));
-        expect(editor.document.getText()).to.be.eq(rawKitsFromFile);
-      } else {
-      }
-    }).timeout(10000);
-  });
-
-  test('KitManager tests event on change of active kit', async () => {
-    let stateMock = sinon.createStubInstance(state.StateManager);
-    let storedActivatedKitName: string = '';
-    sinon.stub(stateMock, 'activeKitName')
-        .get(function() { return null; })
-        .set(function(kit) { storedActivatedKitName = kit; });
-    const km = new kit.KitManager(stateMock, getTestResourceFilePath('test_kit.json'));
-    await km.initialize();
-    // Check that each time we change the kit, it fires a signal
-    let fired_kit: string|null = null;
-    km.onActiveKitChanged(k => fired_kit = k!.name);
-    for (const kit of km.kits) {
-      const name = kit.name;
-      // Set the kit
-      await km.selectKitByName(name);
-      // Check that we got the signal
-      expect(fired_kit).to.eq(name);
-      // Check that we've saved our change
-      expect(storedActivatedKitName).to.eq(name);
-    }
-    km.dispose();
-  }).timeout(10000);
-
-  test('KitManager test load of kit from test file', async () => {
-    let stateMock = sinon.createStubInstance(state.StateManager);
-    sinon.stub(stateMock, 'activeKitName').get(function() { return null; }).set(function() {});
-    const km = new kit.KitManager(stateMock, getTestResourceFilePath('test_kit.json'));
-
-    await km.initialize();
-
-    expect(km.kits.length).to.eq(6);
-    expect(km.kits[0].name).to.eq("CompilerKit 1");
-    expect(km.kits[1].name).to.eq("CompilerKit 2");
-    expect(km.kits[2].name).to.eq("CompilerKit 3 with PreferedGenerator");
-    expect(km.kits[3].name).to.eq("ToolchainKit 1");
-    expect(km.kits[4].name).to.eq("VSCode Kit 1");
-    expect(km.kits[5].name).to.eq("VSCode Kit 2");
-
-    km.dispose();
-  });
-
-  test('KitManager test selection of last activated kit', async () => {
-    let stateMock = sinon.createStubInstance(state.StateManager);
-
-    sinon.stub(stateMock, 'activeKitName').get(function() { return "ToolchainKit 1"; }).set(function() {});
-    const km = new kit.KitManager(stateMock, getTestResourceFilePath('test_kit.json'));
-
-    await km.initialize();
-
-    expect(km.activeKit).to.be.not.null;
-    if (km.activeKit)
-      expect(km.activeKit.name).to.eq("ToolchainKit 1");
-
-    km.dispose();
-  });
-
-  test('KitManager test selection of a default kit', async () => {
-    let stateMock = sinon.createStubInstance(state.StateManager);
-    sinon.stub(stateMock, 'activeKitName').get(function() { return null; }).set(function() {});
-
-    const km = new kit.KitManager(stateMock, getTestResourceFilePath('test_kit.json'));
-    await km.initialize();
-
-    expect(km.activeKit).to.be.null;
-    km.dispose();
-  });
-
-  test('KitManager test selection of default kit if last activated kit is invalid', async () => {
-    let stateMock = sinon.createStubInstance(state.StateManager);
-    let storedActivatedKitName = "not replaced";
-    sinon.stub(stateMock, 'activeKitName')
-        .get(function() { return "Unknown"; })
-        .set(function(kit) { storedActivatedKitName = kit; });
-
-    const km = new kit.KitManager(stateMock, getTestResourceFilePath('test_kit.json'));
-    await km.initialize();
-
-    expect(km.activeKit).to.be.null;
-    expect(storedActivatedKitName).to.be.null;
-    km.dispose();
-  });
-});
 
 suite('Cache test', async () => {
   test("Read CMake Cache", async function() {
